Rename cortex->prometheus
