+++ cmd/grr/main.go
@@ -4,9 +4,9 @@ import (
 	"log"
 
 	"github.com/go-clix/cli"
-	"github.com/grafana/grizzly/pkg/cortex"
 	"github.com/grafana/grizzly/pkg/grafana"
 	"github.com/grafana/grizzly/pkg/grizzly"
+	"github.com/grafana/grizzly/pkg/prometheus"
 )
 
 // Version is the current version of the grr command.
@@ -54,6 +54,6 @@ func main() {
 func GetProviderRegistry() (grizzly.Registry, error) {
 	registry := grizzly.NewProviderRegistry()
 	registry.RegisterProvider(&grafana.Provider{})
-	registry.RegisterProvider(&cortex.Provider{})
+	registry.RegisterProvider(&prometheus.Provider{})
 	return registry, nil
 }
